pyproject.toml
[project]
name = "transplantapp"
version = "0.1.0"
description = ""
authors = [
    { name = "Patryk Antonik", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-phone-field (>=1.8.1,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
exclude = [
    "docs",
    "tests",
    "migrations",
    "./*/migrations",
    "./**/migrations",
    "./donation/admin.py",
]

[tool.ruff.lint]
select = ["E", "F", "I", "DJ"]

[tool.isort]
profile = "black"
known_first_party = ["donation"]
force_sort_within_sections = true
combine_as_imports = true
lines_after_imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-xdist = "^3.8.0"
pylint = "^3.3.9"
pylint-django = "^2.6.1"
pre-commit = "^4.3.0"
pytest-html = "^4.1.1"
pytest-cov = "^7.0.0"
ruff = "^0.14.0"
black = "^25.9.0"
